requirements.txt
selenium
python-dotenv
google
google-generativeai
fpdf
colorama
pandas

